Move hooks into src/app
