+++ pages/thoughts/[slug].js
@@ -3,8 +3,8 @@ import Link from 'next/link'
 import Page from '../../components/Page'
 import Thought from '../../components/Thought'
 import RelatedContent from '../../components/RelatedContent'
-import useDelayedRelatedContent from '../../src/useDelayedRelatedContent'
-import useLiveBlog from '../../src/useLiveBlog'
+import useDelayedRelatedContent from '../../src/app/hooks/useDelayedRelatedContent'
+import useLiveBlog from '../../src/app/hooks/useLiveBlog'
 import { fetchOneThoughtBySlug } from '../../src/app/factory'
 import withErrorHandling from '../../src/withErrorHandling'
 import withCommonProps from '../../src/withCommonProps'
